Code refactor ♻️
* Create folder to manage User types 🎨
* Move each class into separate file 🚚
