+++ Src/Notion.Client/Models/User/Bot.cs
@@ -1,0 +1,10 @@
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class Bot
+    {
+        [JsonProperty("owner")]
+        public IBotOwner Owner { get; set; }
+    }
+}
