+++ Src/Notion.Client/Models/User/BotOwner/UserOwner.cs
@@ -1,0 +1,12 @@
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class UserOwner : IBotOwner
+    {
+        public string Type { get; set; }
+
+        [JsonProperty("user")]
+        public User User { get; set; }
+    }
+}
