+++ Src/Notion.Client/Models/User/Person.cs
@@ -1,0 +1,10 @@
+using Newtonsoft.Json;
+
+namespace Notion.Client
+{
+    public class Person
+    {
+        [JsonProperty("email")]
+        public string Email { get; set; }
+    }
+}
